requirements.txt
<<<<<<< HEAD
anthropic==0.18.1
python-dotenv==1.0.1
google-generativeai==0.3.2
langchain-core==0.1.32
langchain==0.1.12
langchain-anthropic==0.0.10
langchain-google-genai==0.0.11
faiss-cpu==1.7.4 
=======
anthropic==0.16.0
google-generativeai==0.3.1
python-dotenv==1.0.0
ipaddress==1.0.23
python-nmap==0.7.1
>>>>>>> c078e29 (Add Offensive Security Assessment Tool with AI capabilities)
